Reuse assertAllExportedMethodNoPanic using gotmpl
