+++ metric/noop/noop_test.go
@@ -97,36 +97,6 @@ func TestImplementationNoPanics(t *testing.T) {
 	))
 }
 
-func assertAllExportedMethodNoPanic(rVal reflect.Value, rType reflect.Type) func(*testing.T) {
-	return func(t *testing.T) {
-		for n := 0; n < rType.NumMethod(); n++ {
-			mType := rType.Method(n)
-			if !mType.IsExported() {
-				t.Logf("ignoring unexported %s", mType.Name)
-				continue
-			}
-			m := rVal.MethodByName(mType.Name)
-			if !m.IsValid() {
-				t.Errorf("unknown method for %s: %s", rVal.Type().Name(), mType.Name)
-			}
-
-			numIn := mType.Type.NumIn()
-			if mType.Type.IsVariadic() {
-				numIn--
-			}
-			args := make([]reflect.Value, numIn)
-			for i := range args {
-				aType := mType.Type.In(i)
-				args[i] = reflect.New(aType).Elem()
-			}
-
-			assert.NotPanicsf(t, func() {
-				_ = m.Call(args)
-			}, "%s.%s", rVal.Type().Name(), mType.Name)
-		}
-	}
-}
-
 func TestNewMeterProvider(t *testing.T) {
 	mp := NewMeterProvider()
 	assert.Equal(t, mp, MeterProvider{})
