+++ sdk/log/noop_test.go
@@ -4,11 +4,8 @@
 package log
 
 import (
-	"context"
 	"reflect"
 	"testing"
-
-	"github.com/stretchr/testify/assert"
 )
 
 func TestNoopProcessorNoPanics(t *testing.T) {
@@ -16,40 +13,4 @@ func TestNoopProcessorNoPanics(t *testing.T) {
 		reflect.ValueOf(NewNoopProcessor()),
 		reflect.TypeOf((*Processor)(nil)).Elem(),
 	)(t)
-}
-
-func assertAllExportedMethodNoPanic(rVal reflect.Value, rType reflect.Type) func(*testing.T) {
-	return func(t *testing.T) {
-		for n := 0; n < rType.NumMethod(); n++ {
-			mType := rType.Method(n)
-			if !mType.IsExported() {
-				t.Logf("ignoring unexported %s", mType.Name)
-				continue
-			}
-			m := rVal.MethodByName(mType.Name)
-			if !m.IsValid() {
-				t.Errorf("unknown method for %s: %s", rVal.Type().Name(), mType.Name)
-			}
-
-			numIn := mType.Type.NumIn()
-			if mType.Type.IsVariadic() {
-				numIn--
-			}
-			args := make([]reflect.Value, numIn)
-			ctx := context.Background()
-			for i := range args {
-				aType := mType.Type.In(i)
-				if aType.Name() == "Context" {
-					// Do not panic on a nil context.
-					args[i] = reflect.ValueOf(ctx)
-				} else {
-					args[i] = reflect.New(aType).Elem()
-				}
-			}
-
-			assert.NotPanicsf(t, func() {
-				_ = m.Call(args)
-			}, "%s.%s", rVal.Type().Name(), mType.Name)
-		}
-	}
 }
