+++ trace/noop/noop_test.go
@@ -30,42 +30,6 @@ func TestImplementationNoPanics(t *testing.T) {
 	))
 }
 
-func assertAllExportedMethodNoPanic(rVal reflect.Value, rType reflect.Type) func(*testing.T) {
-	return func(t *testing.T) {
-		for n := 0; n < rType.NumMethod(); n++ {
-			mType := rType.Method(n)
-			if !mType.IsExported() {
-				t.Logf("ignoring unexported %s", mType.Name)
-				continue
-			}
-			m := rVal.MethodByName(mType.Name)
-			if !m.IsValid() {
-				t.Errorf("unknown method for %s: %s", rVal.Type().Name(), mType.Name)
-			}
-
-			numIn := mType.Type.NumIn()
-			if mType.Type.IsVariadic() {
-				numIn--
-			}
-			args := make([]reflect.Value, numIn)
-			ctx := context.Background()
-			for i := range args {
-				aType := mType.Type.In(i)
-				if aType.Name() == "Context" {
-					// Do not panic on a nil context.
-					args[i] = reflect.ValueOf(ctx)
-				} else {
-					args[i] = reflect.New(aType).Elem()
-				}
-			}
-
-			assert.NotPanicsf(t, func() {
-				_ = m.Call(args)
-			}, "%s.%s", rVal.Type().Name(), mType.Name)
-		}
-	}
-}
-
 func TestNewTracerProvider(t *testing.T) {
 	tp := NewTracerProvider()
 	assert.Equal(t, tp, TracerProvider{})
